requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
anyio==3.6.2
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==3.4.8
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.86.0
fastapi-jwt-auth==0.5.0
filelock==3.10.0
h11==0.14.0
huggingface-hub==0.13.2
idna==3.4
motor==3.1.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
passlib==1.7.4
Pillow==9.4.0
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pymongo==4.3.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
schedule==1.1.0
six==1.16.0
sniffio==1.3.0
starlette==0.20.4
tokenizers==0.13.2
torch==1.13.0+cpu
torchvision==0.2.2.post3
tqdm==4.65.0
transformers==4.27.1
typing_extensions==4.4.0
urllib3==1.26.15
uvicorn==0.19.0
